% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dough_height}
\alias{dough_height}
\title{Dough rising time dataset}
\format{
A data frame with 12 rows and 4 columns:
\describe{
\item{loaf_number}{ID of dough}
\item{height}{Height of risen dough, unit unknown}
\item{proofing_time}{Time taken for dough to rise, in min}
\item{proofing_time_fct}{Categorical factor for time taken for dough to rise}
...
}
}
\source{
daewr package - Design and Analysis of Experiments with R, by John Lawson, CRC/Chapman Hall
}
\usage{
dough_height
}
\description{
Dataset that looks at the effect of dough rising time on height risen in an
controlled experiment.
}
\keyword{datasets}
