% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tukey.R
\name{plot_tukey}
\alias{plot_tukey}
\title{Plots the 95\% confidence level to show difference in mean levels of different factors}
\usage{
plot_tukey(tukey_tibble)
}
\arguments{
\item{tukey_tibble}{a tibble df for Tukey post-hoc tests generated using rstatix::tukey_hsd}
}
\value{
A ggplot colored by whether difference is significant based on p-value
}
\description{
Plots the 95\% confidence level to show difference in mean levels of different factors
}
\examples{
library(rstatix)
library(tidyverse)
library(janitor)
iris <- iris \%>\% mutate(Species = factor(Species)) \%>\% clean_names()
m <- aov(sepal_length~species, data = iris)
aov_tukey_results <- m \%>\% tukey_hsd()
plot_tukey(aov_tukey_results)
}
